Reorder controls and remove unused resources
Fix sampling mode area calculation

Reorganize some code.  Refactor to ImageController class.
