+++ .github/workflows/main.yaml
@@ -34,7 +34,7 @@ jobs:
         restore-keys: |
           ${{ runner.os }}-gradle-
     - name: Build with Gradle
-      run: ./gradlew build
+      run: ./gradlew build --parallel --daemon
     - name: Run Unit Tests
       run: ./gradlew test
     - name: Enable KVM
@@ -112,6 +112,11 @@ jobs:
     #   with:
     #     name: color-filter-aar
     #     path: app/build/outputs/aar/*.aar
+    - name: Upload Debug APK artifact
+      uses: actions/upload-artifact@v4
+      with:
+        name: color-filter-debug-apk
+        path: app/build/outputs/apk/debug/app-debug.apk
     - name: Upload APK artifact
       uses: actions/upload-artifact@v4
       with:
