+++ app/src/main/java/com/orbitals/colorfilter/MainActivity.java
@@ -10,23 +10,13 @@ import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.content.res.TypedArray;
 import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.ColorSpace;
-import android.graphics.Matrix;
 import android.graphics.PointF;
-import android.graphics.RectF;
 import android.graphics.SurfaceTexture;
-import android.net.Uri;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.HandlerThread;
 import android.provider.MediaStore;
 import android.util.Log;
 import android.view.MotionEvent;
 import android.view.TextureView;
-import android.view.ViewTreeObserver;
 import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
@@ -34,13 +24,9 @@ import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
-import androidx.exifinterface.media.ExifInterface;
 
 import org.opencv.android.OpenCVLoader;
-import org.opencv.android.Utils;
-import org.opencv.core.Mat;
 
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -57,23 +43,18 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
 
     ArrayList<TermMap> termMaps = new ArrayList<>();
 
-    private ImageFilterProcessor filter;
+    private FilterProcessor filter;
     private CameraController cameraController;
+    private ImageController imageController;
     private UIComponentManager uiManager;
 
     private TextureView textureView;
 
-    private HandlerThread backgroundThread;
     private ActivityResultLauncher<Intent> pickImageLauncher;
     private ActivityResultLauncher<Intent> settingsLauncher;
 
-    private Bitmap loadedImage = null;
-    private Bitmap processedImage = null;
     private boolean isImageMode = false;
     private final PointF lastTouch = new PointF();
-    private final Matrix imageMatrix = new Matrix();
-    private final RectF imageBounds = new RectF();
-    private final float[] matrixValues = new float[9];
 
     // Pinch to zoom and swipe variables
     private float mLastTouchDistance = -1f;
@@ -101,39 +82,20 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
         textureView.setSurfaceTextureListener(this);
         uiManager = new UIComponentManager(this);
 
-        filter = new ImageFilterProcessor();
+        filter = new FilterProcessor();
         // defaults
-        filter.setFilterSettings(0, 14, 100, 100, 1, ImageFilterProcessor.FilterMode.EXCLUDE, termMaps.get(0));
+        filter.setFilterSettings(0, 14, 100, 100, 1, FilterProcessor.FilterMode.EXCLUDE, termMaps.get(0));
         filter.setUseLumSatBCT(false);
         loadSavedSettings();
 
         cameraController = new CameraController(this, textureView, this::checkCameraPermissions, filter, this::updateControls);
+        imageController = new ImageController(this, textureView, filter, this::updateControls);
 
         pickImageLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
-            if (result.getResultCode() == RESULT_OK) {
-                Intent data = result.getData();
-                if (data != null && data.getData() != null) {
-                    Uri imageUri = data.getData();
-                    try {
-                        InputStream inputStream = getContentResolver().openInputStream(imageUri);
-                        BitmapFactory.Options options = new BitmapFactory.Options();
-                        options.inPreferredColorSpace = ColorSpace.get(ColorSpace.Named.SRGB);
-                        loadedImage = BitmapFactory.decodeStream(inputStream, null, options);
-                        int orientation = getOrientation(imageUri);
-                        if (orientation != 0) {
-                            Matrix matrix = new Matrix();
-                            matrix.postRotate(orientation);
-                            loadedImage = Bitmap.createBitmap(loadedImage, 0, 0, loadedImage.getWidth(), loadedImage.getHeight(), matrix, true);
-                        }
-                        isImageMode = true;
-                        setupImageMatrix();
-                        cameraController.closeCamera();
-                        displayLoadedImage();
-                    } catch (Exception e) {
-                        Log.e(TAG, "Error loading image", e);
-                        Toast.makeText(this, getString(R.string.image_load_failed), Toast.LENGTH_SHORT).show();
-                    }
-                }
+            Bitmap newImage = imageController.handleImagePickerResult(result);
+            if (newImage != null) {
+                isImageMode = true;
+                cameraController.closeCamera();
             }
         });
 
@@ -147,19 +109,19 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
                     public void onFilterModeChanged() {
                         switch (filter.getFilterMode()) {
                             case NONE:
-                                filter.setFilterMode(ImageFilterProcessor.FilterMode.INCLUDE);
+                                filter.setFilterMode(FilterProcessor.FilterMode.INCLUDE);
                                 break;
                             case INCLUDE:
-                                filter.setFilterMode(ImageFilterProcessor.FilterMode.EXCLUDE);
+                                filter.setFilterMode(FilterProcessor.FilterMode.EXCLUDE);
                                 break;
                             case EXCLUDE:
-                                filter.setFilterMode(ImageFilterProcessor.FilterMode.BINARY);
+                                filter.setFilterMode(FilterProcessor.FilterMode.BINARY);
                                 break;
                             case BINARY:
-                                filter.setFilterMode(ImageFilterProcessor.FilterMode.SATURATION);
+                                filter.setFilterMode(FilterProcessor.FilterMode.SATURATION);
                                 break;
                             case SATURATION:
-                                filter.setFilterMode(ImageFilterProcessor.FilterMode.NONE);
+                                filter.setFilterMode(FilterProcessor.FilterMode.NONE);
                                 break;
                         }
                         updateControls();
@@ -220,8 +182,7 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
                     public void onCameraSwitch(boolean isImageModeRequested) {
                         if (isImageMode) {
                             isImageMode = false;
-                            loadedImage = null;
-                            processedImage = null;
+                            imageController.clearImage();
                             cameraController.openCamera();
                         } else {
                             cameraController.switchCamera();
@@ -317,7 +278,7 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
         );
 
         if (isImageMode) {
-            displayLoadedImage();
+            imageController.displayLoadedImage();
         }
     }
 
@@ -435,7 +396,7 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
     protected void onResume() {
         super.onResume();
         Log.i(TAG, "onResume");
-        startBackgroundThread();
+        cameraController.startBackgroundThread();
         if (!isImageMode) {
             if (textureView.isAvailable()) {
                 cameraController.openCamera();
@@ -449,27 +410,8 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
     protected void onPause() {
         Log.i(TAG, "onPause");
         cameraController.closeCamera();
-        stopBackgroundThread();
+        cameraController.stopBackgroundThread();
         super.onPause();
-    }
-
-    protected void startBackgroundThread() {
-        backgroundThread = new HandlerThread("Camera Background");
-        backgroundThread.start();
-        cameraController.setBackgroundHandler(new Handler(backgroundThread.getLooper()));
-    }
-
-    protected void stopBackgroundThread() {
-        if (backgroundThread != null) {
-            backgroundThread.quitSafely();
-            try {
-                backgroundThread.join();
-                backgroundThread = null;
-                cameraController.setBackgroundHandler(null);
-            } catch (InterruptedException e) {
-                Log.e(TAG, "InterruptedException", e);
-            }
-        }
     }
 
     @Override
@@ -489,78 +431,12 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
         super.onConfigurationChanged(newConfig);
 
         if (isImageMode) {
-            textureView.getViewTreeObserver().addOnGlobalLayoutListener(new ViewTreeObserver.OnGlobalLayoutListener() {
-                @Override
-                public void onGlobalLayout() {
-                    textureView.getViewTreeObserver().removeOnGlobalLayoutListener(this);
-                    setupImageMatrix();
-                    displayLoadedImage();
-                }
-            });
+            imageController.onConfigurationChanged();
         } else {
             cameraController.reopenCamera();
         }
     }
 
-    private void setupImageMatrix() {
-        float viewWidth = textureView.getWidth();
-        float viewHeight = textureView.getHeight();
-        float imageWidth = loadedImage.getWidth();
-        float imageHeight = loadedImage.getHeight();
-
-        // Calculate scale to fit screen while maintaining aspect ratio
-        float scale = Math.min(viewWidth / imageWidth, viewHeight / imageHeight);
-
-        // Center the image
-        float dx = (viewWidth - imageWidth * scale) / 2;
-        float dy = (viewHeight - imageHeight * scale) / 2;
-
-        imageMatrix.reset();
-        imageMatrix.postScale(scale, scale);
-        imageMatrix.postTranslate(dx, dy);
-
-        // Set bounds for pan limits
-        imageBounds.set(0, 0, imageWidth, imageHeight);
-        imageMatrix.mapRect(imageBounds);
-    }
-
-    private void displayLoadedImage() {
-        displayLoadedImage(false);
-    }
-
-    private void displayLoadedImage(Boolean reuse) {
-        if (loadedImage == null || !isImageMode) return;
-
-        if (processedImage == null || !reuse || filter.getSampleMode()) {
-            Mat inputMat = new Mat();
-            Utils.bitmapToMat(loadedImage, inputMat);
-            if (filter.getSampleMode()) {
-                Mat centerChunk = cameraController.centerOfImage(inputMat, imageMatrix);
-                if (filter.sampleRegion(centerChunk)) {
-                    updateControls();
-                    reuse = false;
-                }
-                centerChunk.release();
-            }
-            if ((processedImage == null || !reuse)) {
-                Mat processedMat = filter.process(inputMat);
-
-                processedImage = Bitmap.createBitmap(processedMat.cols(), processedMat.rows(), Bitmap.Config.ARGB_8888);
-                Utils.matToBitmap(processedMat, processedImage);
-
-                inputMat.release();
-                processedMat.release();
-            }
-        }
-
-        Canvas canvas = textureView.lockCanvas();
-        if (canvas != null) {
-            canvas.drawColor(Color.BLACK);
-            canvas.drawBitmap(processedImage, imageMatrix, null);
-            cameraController.drawSamplingCircle(canvas);
-            textureView.unlockCanvasAndPost(canvas);
-        }
-    }
 
     private boolean handleImageTouch(MotionEvent event) {
         if (!isImageMode) return false;
@@ -575,20 +451,9 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
                     // Handle zoom
                     float newDistance = getDistance(event);
                     if (mLastTouchDistance > 0) {
-                        float scale = newDistance / mLastTouchDistance;
-
-                        imageMatrix.getValues(matrixValues);
-                        float currentScale = matrixValues[Matrix.MSCALE_X];
-
-                        float minScale = Math.min(textureView.getWidth() / loadedImage.getWidth(), textureView.getHeight() / loadedImage.getHeight());
-                        float maxScale = Math.max(2.0f, Math.max(textureView.getWidth() / (float) loadedImage.getWidth(), textureView.getHeight() / (float) loadedImage.getHeight()) * 2);
-
-                        float newScale = Math.min(Math.max(currentScale * scale, minScale), maxScale);
-                        scale = newScale / currentScale;
-
                         float centerX = (event.getX(0) + event.getX(1)) / 2;
                         float centerY = (event.getY(0) + event.getY(1)) / 2;
-                        imageMatrix.postScale(scale, scale, centerX, centerY);
+                        imageController.scaleMatrix(newDistance / mLastTouchDistance, centerX, centerY);
                     }
                     mLastTouchDistance = newDistance;
                     lastTouch.set(-1, -1);
@@ -597,13 +462,11 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
                         // Handle pan
                         float dx = event.getX() - lastTouch.x;
                         float dy = event.getY() - lastTouch.y;
+                        imageController.translateMatrix(dx, dy);
 
-                        imageMatrix.postTranslate(dx, dy);
                     }
                     lastTouch.set(event.getX(), event.getY());
                 }
-                constrainImage();
-                displayLoadedImage(true);
                 return true;
             case MotionEvent.ACTION_UP:
             case MotionEvent.ACTION_POINTER_UP:
@@ -612,61 +475,6 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
                 return true;
         }
         return false;
-    }
-
-    private void constrainImage() {
-        RectF rect = new RectF(0, 0, loadedImage.getWidth(), loadedImage.getHeight());
-        imageMatrix.mapRect(rect);
-
-        float dx = 0, dy = 0;
-        float minScale = Math.min((float) textureView.getWidth() / loadedImage.getWidth(), (float) textureView.getHeight() / loadedImage.getHeight());
-        imageMatrix.getValues(matrixValues);
-        float currentScale = matrixValues[Matrix.MSCALE_X];
-        if (currentScale < minScale) {
-            imageMatrix.postScale(minScale / currentScale, minScale / currentScale, textureView.getWidth() / 2f, textureView.getHeight() / 2f);
-            imageMatrix.mapRect(rect);
-        }
-
-        // Constrain horizontal movement
-        if (rect.width() <= textureView.getWidth()) {
-            dx = (textureView.getWidth() - rect.width()) / 2 - rect.left;
-        } else {
-            if (rect.left > 0) dx = -rect.left;
-            if (rect.right < textureView.getWidth()) dx = textureView.getWidth() - rect.right;
-        }
-
-        // Constrain vertical movement
-        if (rect.height() <= textureView.getHeight()) {
-            dy = (textureView.getHeight() - rect.height()) / 2 - rect.top;
-        } else {
-            if (rect.top > 0) dy = -rect.top;
-            if (rect.bottom < textureView.getHeight()) dy = textureView.getHeight() - rect.bottom;
-        }
-        imageMatrix.postTranslate(dx, dy);
-    }
-
-    private int getOrientation(Uri photoUri) {
-        try {
-            InputStream stream = getContentResolver().openInputStream(photoUri);
-            if (stream == null) {
-                return 0;
-            }
-            ExifInterface ei = new ExifInterface(stream);
-            int orientation = ei.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL);
-
-            switch (orientation) {
-                case ExifInterface.ORIENTATION_ROTATE_90:
-                    return 90;
-                case ExifInterface.ORIENTATION_ROTATE_180:
-                    return 180;
-                case ExifInterface.ORIENTATION_ROTATE_270:
-                    return 270;
-                default:
-                    return 0;
-            }
-        } catch (Exception e) {
-            return 0;
-        }
     }
 
     public Boolean checkCameraPermissions() {
@@ -687,8 +495,8 @@ public class MainActivity extends AppCompatActivity implements TextureView.Surfa
 
         if (includeDefaults) {
             int curTerm = filter.getTerm();
-            int filterModeOrdinal = prefs.getInt(SettingsActivity.KEY_FILTER_MODE, ImageFilterProcessor.FilterMode.EXCLUDE.ordinal());
-            filter.setFilterMode(ImageFilterProcessor.FilterMode.values()[filterModeOrdinal]);
+            int filterModeOrdinal = prefs.getInt(SettingsActivity.KEY_FILTER_MODE, FilterProcessor.FilterMode.EXCLUDE.ordinal());
+            filter.setFilterMode(FilterProcessor.FilterMode.values()[filterModeOrdinal]);
             filter.setHue(prefs.getInt(SettingsActivity.KEY_HUE, filter.getHue()));
             filter.setHueWidth(prefs.getInt(SettingsActivity.KEY_HUE_WIDTH, filter.getHueWidth()));
             filter.setSatThreshold(prefs.getInt(SettingsActivity.KEY_SAT_THRESHOLD, filter.getSatThreshold()));
