+++ app/src/main/res/layout/activity_main.xml
@@ -24,71 +24,10 @@
         android:padding="8dp">
 
         <LinearLayout
-            style="?android:attr/buttonBarStyle"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/switchCameraButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="48dp"
-                android:layout_height="wrap_content"
-                android:contentDescription="@string/camera_button"
-                android:drawableStart="@drawable/baseline_camera_alt_24" />
-            <Button
-                android:id="@+id/loadImageButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="48dp"
-                android:layout_height="wrap_content"
-                android:contentDescription="@string/image_button"
-                android:drawableStart="@drawable/baseline_image_24" />
-
-            <Button
-                android:id="@+id/filterButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="90dp"
-                android:layout_height="wrap_content"
-                android:maxLines="1"
-                android:ellipsize="none"
-                android:scrollHorizontally="false"
-                android:text="@string/filter_button_off"
-                android:textColor="?android:attr/textColorPrimary" />
-
-            <Button
-                android:id="@+id/bctButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="68dp"
-                android:layout_height="wrap_content"
-                android:maxLines="1"
-                android:ellipsize="none"
-                android:scrollHorizontally="false"
-                android:text="@string/term_button_hsv"
-                android:textColor="?android:attr/textColorPrimary" />
-
-            <Button
-                android:id="@+id/sampleButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="48dp"
-                android:layout_height="wrap_content"
-                android:contentDescription="@string/sampling_mode"
-                android:drawableStart="@drawable/sample_button_selector"
-                />
-
-            <Button
-                android:id="@+id/settingsButton"
-                style="?android:attr/buttonBarButtonStyle"
-                android:layout_width="48dp"
-                android:layout_height="wrap_content"
-                android:contentDescription="@string/settings_button"
-                android:drawableStart="@drawable/baseline_settings_24" />
-        </LinearLayout>
-
-        <FrameLayout
             android:id="@+id/hueBCTContainer"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
 
             <LinearLayout
                 android:id="@+id/hueControls"
@@ -153,7 +92,7 @@
                     android:thumb="@drawable/seekbar_thumb"
                     android:thumbOffset="9dp" />
             </LinearLayout>
-        </FrameLayout>
+        </LinearLayout>
 
         <LinearLayout
             android:id="@+id/satLumControls"
@@ -193,6 +132,70 @@
                 android:thumb="@drawable/seekbar_thumb"
                 android:thumbOffset="9dp" />
         </LinearLayout>
+
+        <LinearLayout
+            style="?android:attr/buttonBarStyle"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:orientation="horizontal"
+            android:paddingTop="8dp">
+
+            <Button
+                android:id="@+id/switchCameraButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:contentDescription="@string/camera_button"
+                android:drawableStart="@drawable/baseline_camera_alt_24" />
+
+            <Button
+                android:id="@+id/loadImageButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:contentDescription="@string/image_button"
+                android:drawableStart="@drawable/baseline_image_24" />
+
+            <Button
+                android:id="@+id/filterButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="90dp"
+                android:layout_height="wrap_content"
+                android:ellipsize="none"
+                android:maxLines="1"
+                android:scrollHorizontally="false"
+                android:text="@string/filter_button_off"
+                android:textColor="?android:attr/textColorPrimary" />
+
+            <Button
+                android:id="@+id/bctButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="68dp"
+                android:layout_height="wrap_content"
+                android:ellipsize="none"
+                android:maxLines="1"
+                android:scrollHorizontally="false"
+                android:text="@string/term_button_hsv"
+                android:textColor="?android:attr/textColorPrimary" />
+
+            <Button
+                android:id="@+id/sampleButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:contentDescription="@string/sampling_mode"
+                android:drawableStart="@drawable/sample_button_selector" />
+
+            <Button
+                android:id="@+id/settingsButton"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:contentDescription="@string/settings_button"
+                android:drawableStart="@drawable/baseline_settings_24" />
+        </LinearLayout>
+
     </LinearLayout>
 
 </RelativeLayout>
