+++ app/src/main/res/layout/activity_settings.xml
@@ -83,20 +83,6 @@
                 android:textSize="12sp" />
 
             <Button
-                android:id="@+id/setDefaultsButton"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/set_defaults_button" />
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="16dp"
-                android:layout_marginBottom="16dp"
-                android:text="@string/set_defaults_description"
-                android:textSize="12sp" />
-
-            <Button
                 android:id="@+id/loadDefaultsButton"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -108,6 +94,20 @@
                 android:layout_marginStart="16dp"
                 android:layout_marginBottom="16dp"
                 android:text="@string/load_defaults_description"
+                android:textSize="12sp" />
+
+            <Button
+                android:id="@+id/setDefaultsButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/set_defaults_button" />
+
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="16dp"
+                android:layout_marginBottom="16dp"
+                android:text="@string/set_defaults_description"
                 android:textSize="12sp" />
 
         </LinearLayout>
